working draft #27 selfHeal + sourcer.updateSettings
